Removing the Manager dependency in ToManyField.dehydrate()
In fields.py:726 ToManyField.dehydrate() assumes the_m2m to be a Manager. At
that point, it only needs a list of objects to dehydrate. Thus, IMHO, it's
better and more general to be able to handle a list or tuple of objects.

This way, we can manipulate the list of objects to return in the model layer,
in case we need it.